+++ zcash_client_backend/src/data_api/testing.rs
@@ -59,6 +59,8 @@ use super::{
     TransactionStatus, WalletCommitmentTrees, WalletRead, WalletSummary, WalletTest, WalletWrite,
     Zip32Derivation, SAPLING_SHARD_HEIGHT,
 };
+#[cfg(feature = "transparent-inputs")]
+use crate::data_api::Balance;
 use crate::{
     data_api::{wallet::TargetHeight, TargetValue},
     fees::{
@@ -2682,8 +2684,9 @@ impl WalletRead for MockWalletDb {
     fn get_transparent_balances(
         &self,
         _account: Self::AccountId,
-        _max_height: BlockHeight,
-    ) -> Result<HashMap<TransparentAddress, Zatoshis>, Self::Error> {
+        _target_height: TargetHeight,
+        _confirmations_policy: ConfirmationsPolicy,
+    ) -> Result<HashMap<TransparentAddress, Balance>, Self::Error> {
         Ok(HashMap::new())
     }