+++ zcash_client_backend/src/data_api/testing/transparent.rs
@@ -1,20 +1,5 @@
-use std::collections::BTreeMap;
-
-use crate::{
-    data_api::{
-        testing::{
-            AddressType, DataStoreFactory, ShieldedProtocol, TestBuilder, TestCache, TestState,
-        },
-        wallet::{
-            decrypt_and_store_transaction, input_selection::GreedyInputSelector,
-            ConfirmationsPolicy,
-        },
-        Account as _, Balance, InputSource, WalletRead, WalletWrite,
-    },
-    fees::{standard, DustOutputPolicy, StandardFeeRule},
-    wallet::WalletTransparentOutput,
-};
 use assert_matches::assert_matches;
+use std::collections::BTreeMap;
 
 use ::transparent::{
     address::TransparentAddress,
@@ -26,6 +11,20 @@ use zcash_primitives::block::BlockHash;
 use zcash_protocol::{local_consensus::LocalNetwork, value::Zatoshis};
 
 use super::TestAccount;
+use crate::{
+    data_api::{
+        testing::{
+            AddressType, DataStoreFactory, ShieldedProtocol, TestBuilder, TestCache, TestState,
+        },
+        wallet::{
+            decrypt_and_store_transaction, input_selection::GreedyInputSelector,
+            ConfirmationsPolicy, TargetHeight,
+        },
+        Account as _, Balance, InputSource, WalletRead, WalletWrite,
+    },
+    fees::{standard, DustOutputPolicy, StandardFeeRule},
+    wallet::WalletTransparentOutput,
+};
 
 /// Checks whether the transparent balance of the given test `account` is as `expected`
 /// considering the `confirmations_policy`.
@@ -52,19 +51,19 @@ fn check_balance<DSF>(
     assert_eq!(balance.unshielded_balance(), expected);
 
     // Check the older APIs for consistency.
-    let mempool_height = st.wallet().chain_height().unwrap().unwrap() + 1;
+    let target_height = TargetHeight::from(st.wallet().chain_height().unwrap().unwrap() + 1);
     assert_eq!(
         st.wallet()
-            .get_transparent_balances(account.id(), mempool_height)
+            .get_transparent_balances(account.id(), target_height, confirmations_policy)
             .unwrap()
             .get(taddr)
             .cloned()
-            .unwrap_or(Zatoshis::ZERO),
+            .map_or(Zatoshis::ZERO, |b| b.spendable_value()),
         expected.total(),
     );
     assert_eq!(
         st.wallet()
-            .get_spendable_transparent_outputs(taddr, mempool_height.into(), confirmations_policy)
+            .get_spendable_transparent_outputs(taddr, target_height, confirmations_policy)
             .unwrap()
             .into_iter()
             .map(|utxo| utxo.value())
@@ -97,7 +96,11 @@ where
 
     let bal_absent = st
         .wallet()
-        .get_transparent_balances(account_id, height_1)
+        .get_transparent_balances(
+            account_id,
+            TargetHeight::from(height_1 + 1),
+            ConfirmationsPolicy::MIN,
+        )
         .unwrap();
     assert!(bal_absent.is_empty());
 
@@ -119,7 +122,7 @@ where
     assert_matches!(
         st.wallet().get_spendable_transparent_outputs(
             taddr,
-            height_1.into(),
+            TargetHeight::from(height_1 + 1),
             ConfirmationsPolicy::MIN
         ).as_deref(),
         Ok([ret]) if (ret.outpoint(), ret.txout(), ret.mined_height()) == (utxo.outpoint(), utxo.txout(), Some(height_1))
@@ -140,7 +143,11 @@ where
     // Confirm that we no longer see any unspent outputs as of `height_1`.
     assert_matches!(
         st.wallet()
-            .get_spendable_transparent_outputs(taddr, height_1.into(), ConfirmationsPolicy::MIN)
+            .get_spendable_transparent_outputs(
+                taddr,
+                TargetHeight::from(height_1 + 1),
+                ConfirmationsPolicy::MIN
+            )
             .as_deref(),
         Ok(&[])
     );
@@ -154,14 +161,18 @@ where
     // If we include `height_2` then the output is returned.
     assert_matches!(
         st.wallet()
-            .get_spendable_transparent_outputs(taddr, height_2.into(), ConfirmationsPolicy::MIN)
+            .get_spendable_transparent_outputs(taddr, TargetHeight::from(height_2 + 1), ConfirmationsPolicy::MIN)
             .as_deref(),
         Ok([ret]) if (ret.outpoint(), ret.txout(), ret.mined_height()) == (utxo.outpoint(), utxo.txout(), Some(height_2))
     );
 
     assert_matches!(
-        st.wallet().get_transparent_balances(account_id, height_2),
-        Ok(h) if h.get(taddr) == Some(&value)
+        st.wallet().get_transparent_balances(
+            account_id,
+            TargetHeight::from(height_2 + 1),
+            ConfirmationsPolicy::MIN
+        ),
+        Ok(h) if h.get(taddr).map(|b| b.spendable_value()) == Some(value)
     );
 }
 
@@ -386,7 +397,7 @@ where
         &st,
         &account,
         taddr,
-        ConfirmationsPolicy::new_symmetrical_unchecked(2),
+        ConfirmationsPolicy::new_symmetrical_unchecked(2, false),
         &not_confirmed_yet_value,
     );
 
@@ -405,7 +416,7 @@ where
         &st,
         &account,
         taddr,
-        ConfirmationsPolicy::new_symmetrical_unchecked(2),
+        ConfirmationsPolicy::new_symmetrical_unchecked(2, true),
         &zero_or_one_conf_value,
     );
 }